fix: move R1CS compiler to manta-crypto
