+++ manta-crypto/Cargo.toml
@@ -32,6 +32,8 @@ arkworks = [
     "ark-r1cs-std",
     "ark-relations",
     "ark-serialize",
+    "ark-snark",
+    "ark-std"
 ]
 
 # Dalek Cryptography Backend
@@ -45,7 +47,7 @@ serde = [
     "ed25519-dalek?/serde",
     "manta-util/serde-alloc",
     "manta-util/serde-array",
-    "rand_chacha?/serde1",
+    "rand_chacha?/serde1"
 ]
 
 # Standard Library
@@ -56,12 +58,14 @@ std = [
     "ark-ed-on-bls12-381?/std",
     "ark-ed-on-bn254?/std",
     "ark-ff?/std",
+    "ark-groth16?/std",
     "ark-r1cs-std?/std",
     "ark-relations?/std",
     "ark-serialize?/std",
+    "ark-std?/std",
     "manta-util/std",
     "rand?/std",
-    "rand_chacha?/std",
+    "rand_chacha?/std"
 ]
 
 # Testing Frameworks
@@ -74,9 +78,12 @@ ark-ec = { version = "0.3.0", optional = true, default-features = false }
 ark-ed-on-bls12-381 = { version = "0.3.0", optional = true, default-features = false, features = ["r1cs"] }
 ark-ed-on-bn254 = { version = "0.3.0", optional = true, default-features = false, features = ["r1cs"] }
 ark-ff = { version = "0.3.0", optional = true, default-features = false }
+ark-groth16 = { version = "0.3.0", optional = true, default-features = false }
 ark-r1cs-std = { version = "0.3.1", optional = true, default-features = false }
 ark-relations = { version = "0.3.0", optional = true, default-features = false }
 ark-serialize = { version = "0.3.0", optional = true, default-features = false, features = ["derive"] }
+ark-snark = { version = "0.3.0", optional = true, default-features = false }
+ark-std = { version = "0.3.0", optional = true, default-features = false }
 derivative = { version = "2.2.0", default-features = false, features = ["use_core"] }
 ed25519-dalek = { version = "1.0.1", optional = true, default-features = false, features = ["u64_backend"] }
 manta-util = { path = "../manta-util", default-features = false, features = ["alloc"] }
