+++ manta-crypto/src/arkworks/constraint/mod.rs
@@ -14,14 +14,12 @@
 // You should have received a copy of the GNU General Public License
 // along with manta-rs.  If not, see <http://www.gnu.org/licenses/>.
 
-//! Arkworks Constraint System and Proof System Implementations
+//! Arkworks Constraint-System Backends
 
-use alloc::vec::Vec;
-use core::iter::{self, Extend};
-use manta_crypto::{
-    algebra,
+use crate::{
     arkworks::{
-        ff::{Field, FpParameters, PrimeField, ToConstraintField},
+        constraint::fp::Fp,
+        ff::{FpParameters, PrimeField},
         r1cs_std::{
             alloc::AllocVar, eq::EqGadget, fields::FieldVar, select::CondSelectGadget, ToBitsGadget,
         },
@@ -33,335 +31,26 @@ use manta_crypto::{
             },
         },
     },
-    constraint::{
-        measure::{Count, Measure},
-        Input, ProofSystem,
-    },
+    constraint::measure::{Count, Measure},
     eclair::{
         self,
         alloc::{
             mode::{self, Public, Secret},
             Constant, Variable,
         },
-        bool::{Assert, Bool, ConditionalSelect, ConditionalSwap},
+        bool::{Assert, ConditionalSelect, ConditionalSwap},
         num::{AssertWithinBitRange, Zero},
-        ops::{Add, BitAnd, BitOr},
+        ops::Add,
         Has, NonNative,
     },
-    rand::{RngCore, Sample},
-};
-use manta_util::{
-    byte_count,
-    codec::{Decode, DecodeError, Encode, Read, Write},
-    SizeLimit,
 };
 
-#[cfg(feature = "serde")]
-use manta_util::serde::{Deserialize, Serialize, Serializer};
-
-pub use manta_crypto::arkworks::{
+pub use crate::arkworks::{
     r1cs_std::{bits::boolean::Boolean, fields::fp::FpVar},
     relations::r1cs::SynthesisError,
 };
 
-pub mod codec;
-pub mod pairing;
-
-#[cfg(feature = "groth16")]
-#[cfg_attr(doc_cfg, doc(cfg(feature = "groth16")))]
-pub mod groth16;
-
-/// Field Element
-#[cfg_attr(
-    feature = "serde",
-    derive(Deserialize, Serialize),
-    serde(
-        bound(deserialize = "", serialize = ""),
-        crate = "manta_util::serde",
-        deny_unknown_fields,
-        try_from = "Vec<u8>"
-    )
-)]
-#[derive(derivative::Derivative)]
-#[derivative(Clone, Copy, Debug, Default, Eq, Hash, Ord, PartialEq, PartialOrd)]
-pub struct Fp<F>(
-    /// Field Element
-    #[cfg_attr(
-        feature = "serde",
-        serde(serialize_with = "serialize_field_element::<F, _>")
-    )]
-    pub F,
-)
-where
-    F: Field;
-
-impl<F> From<u128> for Fp<F>
-where
-    F: Field,
-{
-    #[inline]
-    fn from(value: u128) -> Self {
-        Self(value.into())
-    }
-}
-
-impl<F> ToConstraintField<F> for Fp<F>
-where
-    F: PrimeField,
-{
-    #[inline]
-    fn to_field_elements(&self) -> Option<Vec<F>> {
-        self.0.to_field_elements()
-    }
-}
-
-impl<F, P> Input<P> for Fp<F>
-where
-    F: Field,
-    P: ProofSystem + ?Sized,
-    P::Input: Extend<F>,
-{
-    #[inline]
-    fn extend(&self, input: &mut P::Input) {
-        input.extend(iter::once(self.0))
-    }
-}
-
-impl<F> Decode for Fp<F>
-where
-    F: Field,
-{
-    type Error = codec::SerializationError;
-
-    #[inline]
-    fn decode<R>(reader: R) -> Result<Self, DecodeError<R::Error, Self::Error>>
-    where
-        R: Read,
-    {
-        let mut reader = codec::ArkReader::new(reader);
-        match F::deserialize(&mut reader) {
-            Ok(value) => reader
-                .finish()
-                .map(move |_| Self(value))
-                .map_err(DecodeError::Read),
-            Err(err) => Err(DecodeError::Decode(err)),
-        }
-    }
-}
-
-impl<F> Encode for Fp<F>
-where
-    F: Field,
-{
-    #[inline]
-    fn encode<W>(&self, writer: W) -> Result<(), W::Error>
-    where
-        W: Write,
-    {
-        let mut writer = codec::ArkWriter::new(writer);
-        let _ = self.0.serialize(&mut writer);
-        writer.finish().map(move |_| ())
-    }
-}
-
-#[cfg(feature = "scale")]
-#[cfg_attr(doc_cfg, doc(cfg(feature = "scale")))]
-impl<F> scale_codec::Decode for Fp<F>
-where
-    F: Field,
-{
-    #[inline]
-    fn decode<I>(input: &mut I) -> Result<Self, scale_codec::Error>
-    where
-        I: scale_codec::Input,
-    {
-        Ok(Self(
-            F::deserialize(codec::ScaleCodecReader(input)).map_err(|_| "Deserialization Error")?,
-        ))
-    }
-}
-
-#[cfg(feature = "scale")]
-#[cfg_attr(doc_cfg, doc(cfg(feature = "scale")))]
-impl<F> scale_codec::Encode for Fp<F>
-where
-    F: Field,
-{
-    #[inline]
-    fn using_encoded<R, Encoder>(&self, f: Encoder) -> R
-    where
-        Encoder: FnOnce(&[u8]) -> R,
-    {
-        f(&field_element_as_bytes::<F>(&self.0))
-    }
-}
-
-#[cfg(feature = "scale")]
-#[cfg_attr(doc_cfg, doc(cfg(feature = "scale")))]
-impl<F> scale_codec::EncodeLike for Fp<F> where F: Field {}
-
-#[cfg(feature = "scale")]
-#[cfg_attr(doc_cfg, doc(cfg(feature = "scale")))]
-impl<F> scale_codec::MaxEncodedLen for Fp<F>
-where
-    F: Field,
-{
-    #[inline]
-    fn max_encoded_len() -> usize {
-        byte_count(
-            <<F::BasePrimeField as PrimeField>::Params as FpParameters>::MODULUS_BITS
-                * (F::extension_degree() as u32),
-        ) as usize
-    }
-}
-
-#[cfg(feature = "scale")]
-#[cfg_attr(doc_cfg, doc(cfg(feature = "scale")))]
-impl<F> scale_info::TypeInfo for Fp<F>
-where
-    F: Field,
-{
-    type Identity = [u8];
-
-    #[inline]
-    fn type_info() -> scale_info::Type {
-        Self::Identity::type_info()
-    }
-}
-
-impl<F> eclair::cmp::PartialEq<Self> for Fp<F>
-where
-    F: Field,
-{
-    #[inline]
-    fn eq(&self, rhs: &Self, _: &mut ()) -> bool {
-        PartialEq::eq(self, rhs)
-    }
-}
-
-impl<F> eclair::num::Zero for Fp<F>
-where
-    F: Field,
-{
-    type Verification = bool;
-
-    #[inline]
-    fn zero(_: &mut ()) -> Self {
-        Self(F::zero())
-    }
-
-    #[inline]
-    fn is_zero(&self, _: &mut ()) -> Self::Verification {
-        self.0.is_zero()
-    }
-}
-
-impl<F> eclair::num::One for Fp<F>
-where
-    F: Field,
-{
-    type Verification = bool;
-
-    #[inline]
-    fn one(_: &mut ()) -> Self {
-        Self(F::one())
-    }
-
-    #[inline]
-    fn is_one(&self, _: &mut ()) -> Self::Verification {
-        self.0.is_one()
-    }
-}
-
-impl<F> ConditionalSelect for Fp<F>
-where
-    F: Field,
-{
-    #[inline]
-    fn select(bit: &Bool, true_value: &Self, false_value: &Self, _: &mut ()) -> Self {
-        if *bit {
-            *true_value
-        } else {
-            *false_value
-        }
-    }
-}
-
-impl<F> Sample for Fp<F>
-where
-    F: Field,
-{
-    #[inline]
-    fn sample<R>(_: (), rng: &mut R) -> Self
-    where
-        R: RngCore + ?Sized,
-    {
-        Self(F::rand(rng))
-    }
-}
-
-impl<F> algebra::Group for Fp<F>
-where
-    F: Field,
-{
-    #[inline]
-    fn add(&self, rhs: &Self, _: &mut ()) -> Self {
-        Self(self.0 + rhs.0)
-    }
-}
-
-impl<F> algebra::Ring for Fp<F>
-where
-    F: Field,
-{
-    #[inline]
-    fn mul(&self, rhs: &Self, _: &mut ()) -> Self {
-        Self(self.0 * rhs.0)
-    }
-}
-
-impl<F> SizeLimit for Fp<F>
-where
-    F: PrimeField,
-{
-    const SIZE: usize = byte_count(<F::Params as FpParameters>::MODULUS_BITS) as usize;
-}
-
-impl<F> TryFrom<Vec<u8>> for Fp<F>
-where
-    F: Field,
-{
-    type Error = codec::SerializationError;
-
-    #[inline]
-    fn try_from(bytes: Vec<u8>) -> Result<Self, Self::Error> {
-        F::deserialize(&mut bytes.as_slice()).map(Self)
-    }
-}
-
-/// Converts `element` into its canonical byte-representation.
-#[inline]
-pub fn field_element_as_bytes<F>(element: &F) -> Vec<u8>
-where
-    F: Field,
-{
-    let mut buffer = Vec::new();
-    element
-        .serialize(&mut buffer)
-        .expect("Serialization is not allowed to fail.");
-    buffer
-}
-
-/// Uses `serializer` to serialize `element`.
-#[cfg(feature = "serde")]
-#[inline]
-fn serialize_field_element<F, S>(element: &F, serializer: S) -> Result<S::Ok, S::Error>
-where
-    F: Field,
-    S: Serializer,
-{
-    serializer.serialize_bytes(&field_element_as_bytes(element))
-}
+pub mod fp;
 
 /// Synthesis Result
 pub type SynthesisResult<T = ()> = Result<T, SynthesisError>;
@@ -370,7 +59,7 @@ pub type SynthesisResult<T = ()> = Result<T, SynthesisError>;
 ///
 /// # Warning
 ///
-/// This does not work for all variable assignments! For some assignemnts, the variable inherits
+/// This does not work for all variable assignments! For some assignments, the variable inherits
 /// some structure from its input, like its length or number of bits, which are only known at
 /// run-time. For those cases, some mocking is required and this function can not be used directly.
 #[inline]
@@ -384,44 +73,70 @@ pub fn full<T>(value: T) -> impl FnOnce() -> SynthesisResult<T> {
     move || Ok(value)
 }
 
-/// Arkworks Rank-1 Constraint System
-pub struct R1CS<F>
+/// Conditionally select from `lhs` and `rhs` depending on the value of `bit`.
+#[inline]
+pub fn conditionally_select<F, T>(bit: &Boolean<F>, lhs: &T, rhs: &T) -> T
 where
     F: PrimeField,
+    T: CondSelectGadget<F>,
 {
-    /// Constraint System
-    pub(crate) cs: ConstraintSystemRef<F>,
+    CondSelectGadget::conditionally_select(bit, lhs, rhs)
+        .expect("Conditionally selecting from two values is not allowed to fail.")
 }
+
+/// Arkworks Rank-1 Constraint System
+#[derive(derivative::Derivative)]
+#[derivative(Clone, Debug)]
+pub struct R1CS<F>(ConstraintSystemRef<F>)
+where
+    F: PrimeField;
 
 impl<F> R1CS<F>
 where
     F: PrimeField,
 {
+    /// Builds a new [`R1CS`] constraint system from `constraint_system` without checking its
+    /// optimization goal or synthesis mode.
+    #[inline]
+    pub fn new_unchecked(constraint_system: ConstraintSystemRef<F>) -> Self {
+        Self(constraint_system)
+    }
+
     /// Constructs a new constraint system which is ready for unknown variables.
     #[inline]
     pub fn for_contexts() -> Self {
         // FIXME: This might not be the right setup for all proof systems.
-        let cs = ConstraintSystem::new_ref();
-        cs.set_optimization_goal(OptimizationGoal::Constraints);
-        cs.set_mode(SynthesisMode::Setup);
-        Self { cs }
+        let constraint_system = ConstraintSystem::new_ref();
+        constraint_system.set_optimization_goal(OptimizationGoal::Constraints);
+        constraint_system.set_mode(SynthesisMode::Setup);
+        Self::new_unchecked(constraint_system)
     }
 
     /// Constructs a new constraint system which is ready for known variables.
     #[inline]
     pub fn for_proofs() -> Self {
         // FIXME: This might not be the right setup for all proof systems.
-        let cs = ConstraintSystem::new_ref();
-        cs.set_optimization_goal(OptimizationGoal::Constraints);
-        Self { cs }
+        let constraint_system = ConstraintSystem::new_ref();
+        constraint_system.set_optimization_goal(OptimizationGoal::Constraints);
+        Self::new_unchecked(constraint_system)
     }
 
     /// Check if all constraints are satisfied.
     #[inline]
     pub fn is_satisfied(&self) -> bool {
-        self.cs
+        self.0
             .is_satisfied()
-            .expect("is_satisfied is not allowed to fail")
+            .expect("Checking circuit satisfaction is not allowed to fail.")
+    }
+}
+
+impl<F> AsRef<ConstraintSystemRef<F>> for R1CS<F>
+where
+    F: PrimeField,
+{
+    #[inline]
+    fn as_ref(&self) -> &ConstraintSystemRef<F> {
+        &self.0
     }
 }
 
@@ -445,6 +160,121 @@ where
     }
 }
 
+impl<F> Count<mode::Constant> for R1CS<F> where F: PrimeField {}
+
+impl<F> Count<Public> for R1CS<F>
+where
+    F: PrimeField,
+{
+    #[inline]
+    fn count(&self) -> Option<usize> {
+        Some(self.0.num_instance_variables())
+    }
+}
+
+impl<F> Count<Secret> for R1CS<F>
+where
+    F: PrimeField,
+{
+    #[inline]
+    fn count(&self) -> Option<usize> {
+        Some(self.0.num_witness_variables())
+    }
+}
+
+impl<F> Measure for R1CS<F>
+where
+    F: PrimeField,
+{
+    #[inline]
+    fn constraint_count(&self) -> usize {
+        self.0.num_constraints()
+    }
+}
+
+impl<F> ConstraintSynthesizer<F> for R1CS<F>
+where
+    F: PrimeField,
+{
+    /// Generates constraints for `self` by copying them into `cs`. This method is necessary to hook
+    /// into the proof system traits defined in `arkworks`.
+    #[inline]
+    fn generate_constraints(self, cs: ConstraintSystemRef<F>) -> SynthesisResult {
+        let precomputed_cs = self
+            .0
+            .into_inner()
+            .expect("We own this constraint system so we can consume it.");
+        let mut target_cs = cs
+            .borrow_mut()
+            .expect("This is given to us to mutate so it can't be borrowed by anyone else.");
+        *target_cs = precomputed_cs;
+        Ok(())
+    }
+}
+
+impl<F> Constant<R1CS<F>> for Boolean<F>
+where
+    F: PrimeField,
+{
+    type Type = bool;
+
+    #[inline]
+    fn new_constant(this: &Self::Type, compiler: &mut R1CS<F>) -> Self {
+        AllocVar::new_constant(ns!(compiler.0, "boolean constant"), this)
+            .expect("Variable allocation is not allowed to fail.")
+    }
+}
+
+impl<F> Variable<Public, R1CS<F>> for Boolean<F>
+where
+    F: PrimeField,
+{
+    type Type = bool;
+
+    #[inline]
+    fn new_known(this: &Self::Type, compiler: &mut R1CS<F>) -> Self {
+        Self::new_input(ns!(compiler.0, "boolean public input"), full(this))
+            .expect("Variable allocation is not allowed to fail.")
+    }
+
+    #[inline]
+    fn new_unknown(compiler: &mut R1CS<F>) -> Self {
+        Self::new_input(ns!(compiler.0, "boolean public input"), empty::<bool>)
+            .expect("Variable allocation is not allowed to fail.")
+    }
+}
+
+impl<F> Variable<Secret, R1CS<F>> for Boolean<F>
+where
+    F: PrimeField,
+{
+    type Type = bool;
+
+    #[inline]
+    fn new_known(this: &Self::Type, compiler: &mut R1CS<F>) -> Self {
+        Self::new_witness(ns!(compiler.0, "boolean secret witness"), full(this))
+            .expect("Variable allocation is not allowed to fail.")
+    }
+
+    #[inline]
+    fn new_unknown(compiler: &mut R1CS<F>) -> Self {
+        Self::new_witness(ns!(compiler.0, "boolean secret witness"), empty::<bool>)
+            .expect("Variable allocation is not allowed to fail.")
+    }
+}
+
+impl<F> eclair::cmp::PartialEq<Self, R1CS<F>> for Boolean<F>
+where
+    F: PrimeField,
+{
+    #[inline]
+    fn eq(&self, rhs: &Self, compiler: &mut R1CS<F>) -> Boolean<F> {
+        let _ = compiler;
+        self.is_eq(rhs)
+            .expect("Equality checking is not allowed to fail.")
+    }
+}
+
 impl<F, const BITS: usize> AssertWithinBitRange<FpVar<F>, BITS> for R1CS<F>
 where
     F: PrimeField,
@@ -465,147 +295,6 @@ where
     }
 }
 
-impl<F> Count<mode::Constant> for R1CS<F> where F: PrimeField {}
-
-impl<F> Count<Public> for R1CS<F>
-where
-    F: PrimeField,
-{
-    #[inline]
-    fn count(&self) -> Option<usize> {
-        Some(self.cs.num_instance_variables())
-    }
-}
-
-impl<F> Count<Secret> for R1CS<F>
-where
-    F: PrimeField,
-{
-    #[inline]
-    fn count(&self) -> Option<usize> {
-        Some(self.cs.num_witness_variables())
-    }
-}
-
-impl<F> Measure for R1CS<F>
-where
-    F: PrimeField,
-{
-    #[inline]
-    fn constraint_count(&self) -> usize {
-        self.cs.num_constraints()
-    }
-}
-
-impl<F> ConstraintSynthesizer<F> for R1CS<F>
-where
-    F: PrimeField,
-{
-    /// Generates constraints for `self` by copying them into `cs`. This method is necessary to hook
-    /// into the proof system traits defined in `arkworks`.
-    #[inline]
-    fn generate_constraints(self, cs: ConstraintSystemRef<F>) -> SynthesisResult {
-        let precomputed_cs = self
-            .cs
-            .into_inner()
-            .expect("We own this constraint system so we can consume it.");
-        let mut target_cs = cs
-            .borrow_mut()
-            .expect("This is given to us to mutate so it can't be borrowed by anyone else.");
-        *target_cs = precomputed_cs;
-        Ok(())
-    }
-}
-
-impl<F> Constant<R1CS<F>> for Boolean<F>
-where
-    F: PrimeField,
-{
-    type Type = bool;
-
-    #[inline]
-    fn new_constant(this: &Self::Type, compiler: &mut R1CS<F>) -> Self {
-        AllocVar::new_constant(ns!(compiler.cs, "boolean constant"), this)
-            .expect("Variable allocation is not allowed to fail.")
-    }
-}
-
-impl<F> Variable<Public, R1CS<F>> for Boolean<F>
-where
-    F: PrimeField,
-{
-    type Type = bool;
-
-    #[inline]
-    fn new_known(this: &Self::Type, compiler: &mut R1CS<F>) -> Self {
-        Self::new_input(ns!(compiler.cs, "boolean public input"), full(this))
-            .expect("Variable allocation is not allowed to fail.")
-    }
-
-    #[inline]
-    fn new_unknown(compiler: &mut R1CS<F>) -> Self {
-        Self::new_input(ns!(compiler.cs, "boolean public input"), empty::<bool>)
-            .expect("Variable allocation is not allowed to fail.")
-    }
-}
-
-impl<F> Variable<Secret, R1CS<F>> for Boolean<F>
-where
-    F: PrimeField,
-{
-    type Type = bool;
-
-    #[inline]
-    fn new_known(this: &Self::Type, compiler: &mut R1CS<F>) -> Self {
-        Self::new_witness(ns!(compiler.cs, "boolean secret witness"), full(this))
-            .expect("Variable allocation is not allowed to fail.")
-    }
-
-    #[inline]
-    fn new_unknown(compiler: &mut R1CS<F>) -> Self {
-        Self::new_witness(ns!(compiler.cs, "boolean secret witness"), empty::<bool>)
-            .expect("Variable allocation is not allowed to fail.")
-    }
-}
-
-impl<F> eclair::cmp::PartialEq<Self, R1CS<F>> for Boolean<F>
-where
-    F: PrimeField,
-{
-    #[inline]
-    fn eq(&self, rhs: &Self, compiler: &mut R1CS<F>) -> Boolean<F> {
-        let _ = compiler;
-        self.is_eq(rhs)
-            .expect("Equality checking is not allowed to fail.")
-    }
-}
-
-impl<F> BitAnd<Self, R1CS<F>> for Boolean<F>
-where
-    F: PrimeField,
-{
-    type Output = Self;
-
-    #[inline]
-    fn bitand(self, rhs: Self, compiler: &mut R1CS<F>) -> Self::Output {
-        let _ = compiler;
-        self.and(&rhs).expect("Bitwise AND is not allowed to fail.")
-    }
-}
-
-impl<F> BitOr<Self, R1CS<F>> for Boolean<F>
-where
-    F: PrimeField,
-{
-    type Output = Self;
-
-    #[inline]
-    fn bitor(self, rhs: Self, compiler: &mut R1CS<F>) -> Self::Output {
-        let _ = compiler;
-        self.or(&rhs).expect("Bitwise OR is not allowed to fail.")
-    }
-}
-
 impl<F> Constant<R1CS<F>> for FpVar<F>
 where
     F: PrimeField,
@@ -614,7 +303,7 @@ where
 
     #[inline]
     fn new_constant(this: &Self::Type, compiler: &mut R1CS<F>) -> Self {
-        AllocVar::new_constant(ns!(compiler.cs, "field constant"), this.0)
+        AllocVar::new_constant(ns!(compiler.0, "field constant"), this.0)
             .expect("Variable allocation is not allowed to fail.")
     }
 }
@@ -640,13 +329,13 @@ where
 
     #[inline]
     fn new_known(this: &Self::Type, compiler: &mut R1CS<F>) -> Self {
-        Self::new_input(ns!(compiler.cs, "field public input"), full(this.0))
+        Self::new_input(ns!(compiler.0, "field public input"), full(this.0))
             .expect("Variable allocation is not allowed to fail.")
     }
 
     #[inline]
     fn new_unknown(compiler: &mut R1CS<F>) -> Self {
-        Self::new_input(ns!(compiler.cs, "field public input"), empty::<F>)
+        Self::new_input(ns!(compiler.0, "field public input"), empty::<F>)
             .expect("Variable allocation is not allowed to fail.")
     }
 }
@@ -659,13 +348,13 @@ where
 
     #[inline]
     fn new_known(this: &Self::Type, compiler: &mut R1CS<F>) -> Self {
-        Self::new_witness(ns!(compiler.cs, "field secret witness"), full(this.0))
+        Self::new_witness(ns!(compiler.0, "field secret witness"), full(this.0))
             .expect("Variable allocation is not allowed to fail.")
     }
 
     #[inline]
     fn new_unknown(compiler: &mut R1CS<F>) -> Self {
-        Self::new_witness(ns!(compiler.cs, "field secret witness"), empty::<F>)
+        Self::new_witness(ns!(compiler.0, "field secret witness"), empty::<F>)
             .expect("Variable allocation is not allowed to fail.")
     }
 }
@@ -680,17 +369,6 @@ where
         self.is_eq(rhs)
             .expect("Equality checking is not allowed to fail.")
     }
-}
-
-/// Conditionally select from `lhs` and `rhs` depending on the value of `bit`.
-#[inline]
-pub fn conditionally_select<F, T>(bit: &Boolean<F>, lhs: &T, rhs: &T) -> T
-where
-    F: PrimeField,
-    T: CondSelectGadget<F>,
-{
-    CondSelectGadget::conditionally_select(bit, lhs, rhs)
-        .expect("Conditionally selecting from two values is not allowed to fail.")
 }
 
 impl<F> ConditionalSelect<R1CS<F>> for FpVar<F>
@@ -759,12 +437,12 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
-    use core::iter::repeat_with;
-    use manta_crypto::{
+    use crate::{
         arkworks::{bls12_381::Fr, ff::BigInteger},
         eclair::alloc::Allocate,
         rand::{OsRng, Rand},
     };
+    use core::iter::repeat_with;
 
     /// Checks if `assert_within_range` passes when `should_pass` is `true` and fails when
     /// `should_pass` is `false`.
