+++ manta-crypto/src/arkworks/pairing.rs
@@ -18,8 +18,13 @@
 
 use crate::arkworks::{
     ec::{AffineCurve, PairingEngine},
-    ff::PrimeField,
+    ff::{Fp2, PrimeField},
+    serialize::{
+        CanonicalDeserialize, CanonicalSerialize, HasDeserialization, HasSerialization, Read,
+        SerializationError, Write,
+    },
 };
+use alloc::vec::Vec;
 use core::iter;
 
 /// Pairing Configuration
@@ -105,6 +110,177 @@ pub trait PairingEngineExt: PairingEngine {
 }
 
 impl<E> PairingEngineExt for E where E: PairingEngine {}
+
+/// Generates the `G2Prepared` and `G2PreparedRef` structures for serialization compatibility
+/// with arkworks canonical serialization.
+macro_rules! pairing_impl {
+    ($params:ident) => {
+        /// Line Evaluation Coefficients
+        pub type EllCoeff<F> = (F, F, F);
+
+        /// G2 Prepared Point
+        #[derive(derivative::Derivative, CanonicalSerialize, CanonicalDeserialize)]
+        #[derivative(Clone, Default, Debug, Eq, PartialEq)]
+        pub struct G2Prepared<P>
+        where
+            P: $params,
+        {
+            /// Coefficients
+            pub ell_coeffs: Vec<EllCoeff<Fp2<P::Fp2Params>>>,
+
+            /// Infinity Flag
+            pub infinity: bool,
+        }
+
+        impl<P> From<g2::G2Prepared<P>> for G2Prepared<P>
+        where
+            P: $params,
+        {
+            #[inline]
+            fn from(point: g2::G2Prepared<P>) -> Self {
+                Self {
+                    ell_coeffs: point.ell_coeffs,
+                    infinity: point.infinity,
+                }
+            }
+        }
+
+        impl<P> From<G2Prepared<P>> for g2::G2Prepared<P>
+        where
+            P: $params,
+        {
+            #[inline]
+            fn from(point: G2Prepared<P>) -> Self {
+                Self {
+                    ell_coeffs: point.ell_coeffs,
+                    infinity: point.infinity,
+                }
+            }
+        }
+
+        /// G2 Prepared Point Reference
+        #[derive(derivative::Derivative)]
+        #[derivative(Debug, Eq, PartialEq)]
+        pub struct G2PreparedRef<'p, P>(pub &'p g2::G2Prepared<P>)
+        where
+            P: $params;
+
+        impl<'p, P> CanonicalSerialize for G2PreparedRef<'p, P>
+        where
+            P: $params,
+        {
+            #[inline]
+            fn serialize<W>(&self, mut writer: W) -> Result<(), SerializationError>
+            where
+                W: Write,
+            {
+                let g2::G2Prepared {
+                    ell_coeffs,
+                    infinity,
+                } = &self.0;
+                ell_coeffs.serialize(&mut writer)?;
+                infinity.serialize(&mut writer)?;
+                Ok(())
+            }
+
+            #[inline]
+            fn serialized_size(&self) -> usize {
+                let g2::G2Prepared {
+                    ell_coeffs,
+                    infinity,
+                } = &self.0;
+                ell_coeffs.serialized_size() + infinity.serialized_size()
+            }
+
+            #[inline]
+            fn serialize_uncompressed<W>(&self, mut writer: W) -> Result<(), SerializationError>
+            where
+                W: Write,
+            {
+                let g2::G2Prepared {
+                    ell_coeffs,
+                    infinity,
+                } = &self.0;
+                ell_coeffs.serialize_uncompressed(&mut writer)?;
+                infinity.serialize_uncompressed(&mut writer)?;
+                Ok(())
+            }
+
+            #[inline]
+            fn serialize_unchecked<W>(&self, mut writer: W) -> Result<(), SerializationError>
+            where
+                W: Write,
+            {
+                let g2::G2Prepared {
+                    ell_coeffs,
+                    infinity,
+                } = &self.0;
+                ell_coeffs.serialize_unchecked(&mut writer)?;
+                infinity.serialize_unchecked(&mut writer)?;
+                Ok(())
+            }
+
+            #[inline]
+            fn uncompressed_size(&self) -> usize {
+                let g2::G2Prepared {
+                    ell_coeffs,
+                    infinity,
+                } = &self.0;
+                ell_coeffs.uncompressed_size() + infinity.uncompressed_size()
+            }
+        }
+
+        impl<'p, P> From<&'p g2::G2Prepared<P>> for G2PreparedRef<'p, P>
+        where
+            P: $params,
+        {
+            #[inline]
+            fn from(point: &'p g2::G2Prepared<P>) -> Self {
+                Self(point)
+            }
+        }
+
+        impl<'p, P> From<G2PreparedRef<'p, P>> for &'p g2::G2Prepared<P>
+        where
+            P: $params,
+        {
+            #[inline]
+            fn from(point: G2PreparedRef<'p, P>) -> Self {
+                point.0
+            }
+        }
+
+        impl<'p, P> HasSerialization<'p> for g2::G2Prepared<P>
+        where
+            P: $params,
+        {
+            type Serialize = G2PreparedRef<'p, P>;
+        }
+
+        impl<P> HasDeserialization for g2::G2Prepared<P>
+        where
+            P: $params,
+        {
+            type Deserialize = G2Prepared<P>;
+        }
+    };
+}
+
+/// BLS12 Utilities
+pub mod bls12 {
+    use super::*;
+    use crate::arkworks::ec::models::bls12::{g2, Bls12Parameters};
+
+    pairing_impl!(Bls12Parameters);
+}
+
+/// Bn254 Utilities
+pub mod bn254 {
+    use super::*;
+    use crate::arkworks::ec::models::bn::{g2, BnParameters};
+
+    pairing_impl!(BnParameters);
+}
 
 /// Testing Framework
 #[cfg(any(feature = "test", test))]
