+++ manta-pay/Cargo.toml
@@ -39,7 +39,6 @@ required-features = ["clap", "groth16", "simulation"]
 [features]
 # Enable Arkworks Backend
 arkworks = [
-    "ark-std",
     "manta-crypto/ark-bls12-381",
     "manta-crypto/ark-ed-on-bls12-381",
     "manta-crypto/arkworks",
@@ -49,13 +48,13 @@ arkworks = [
 download = ["manta-parameters/download", "std"]
 
 # Enable Groth16 ZKP System
-groth16 = ["ark-groth16", "ark-snark", "arkworks"]
+groth16 = [
+    "arkworks",
+    "manta-crypto/ark-groth16",
+]
 
-# SCALE Codec and Type Info
+# SCALE Codec and Type Info Support
 scale = ["scale-codec", "scale-info"]
-
-# SCALE Codec and Type Info with the Standard Library Enabled
-scale-std = ["scale", "scale-codec/std", "scale-info/std", "std"]
 
 # Serde
 serde = ["manta-accounting/serde", "manta-crypto/serde"]
@@ -100,9 +99,6 @@ network = []
 
 [dependencies]
 aes-gcm = { version = "0.9.4", default-features = false, features = ["aes", "alloc"] }
-ark-groth16 = { version = "0.3.0", optional = true, default-features = false }
-ark-snark = { version = "0.3.0", optional = true, default-features = false }
-ark-std = { version = "0.3.0", optional = true, default-features = false }
 bip0039 = { version = "0.10.1", optional = true, default-features = false }
 bip32 = { version = "0.3.0", optional = true, default-features = false, features = ["bip39", "secp256k1"] }
 blake2 = { version = "0.10.4", default-features = false }
@@ -126,4 +122,4 @@ ws_stream_wasm = { version = "0.7.3", optional = true, default-features = false 
 
 [dev-dependencies]
 manta-crypto = { path = "../manta-crypto", default-features = false, features = ["getrandom"] }
-manta-pay = { path = ".", default-features = false, features = ["download", "groth16", "scale", "scale-std", "std", "test"] }
+manta-pay = { path = ".", default-features = false, features = ["download", "groth16", "std", "test"] }
