+++ manta-pay/src/crypto/mod.rs
@@ -16,7 +16,6 @@
 
 //! Manta Pay Cryptographic Primitives Implementations
 
-pub mod constraint;
 pub mod ecc;
 pub mod encryption;
 pub mod key;
