+++ manta-pay/src/crypto/poseidon/compat.rs
@@ -327,9 +327,9 @@ pub type Output<S, const ARITY: usize, COM = ()> =
 #[cfg(feature = "arkworks")]
 #[cfg_attr(doc_cfg, doc(cfg(feature = "arkworks")))]
 pub mod arkworks {
-    use crate::crypto::constraint::arkworks::{Fp, FpVar, R1CS};
     use manta_crypto::{
         arkworks::{
+            constraint::{fp::Fp, FpVar, R1CS},
             ff::{Field, PrimeField},
             r1cs_std::fields::FieldVar,
         },
