+++ manta-pay/src/crypto/poseidon/mds.rs
@@ -264,10 +264,11 @@ where
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::crypto::{constraint::arkworks::Fp, poseidon::matrix::Matrix};
+    use crate::crypto::poseidon::matrix::Matrix;
     use manta_crypto::{
         arkworks::{
             bls12_381::Fr,
+            constraint::fp::Fp,
             ff::{field_new, UniformRand},
         },
         rand::OsRng,
